@@ -1,11 +1,11 @@
 {
     "name": "kukrik/bootstrap-filecontrol",
-    "description": "QCubed v4 plugin of FileControl",
+    "description": "QCubed-4 plugin of FileControl",
     "type": "qcubed-library",
     "keywords": ["qcubed", "bootstrap-filecontrol"],
     "license": "MIT",
     "require": {
-        "php": ">=7.4"
+        "php": ">=8.3"
     },
     "autoload": {
       "psr-4": {"QCubed\\Plugin\\": "src/"}
