@@ -3,45 +3,53 @@ require('qcubed.inc.php');
 
 use QCubed as Q;
 use QCubed\Bootstrap as Bs;
-use QCubed\Project\Control\ControlBase;
+use QCubed\Plugin as Qp;
 use QCubed\Project\Control\FormBase as Form;
+use QCubed\Exception\Caller;
 
 class ExamplesForm extends Form
 {
-    protected $btnSingle;
-    protected $btnMultiple;
-    protected $btnFolder;
-    protected $btnGroup1;
-    protected $btnGroup2;
+    protected Qp\BsFileControl $btnSingle;
+    protected Qp\BsFileControl $btnMultiple;
+    protected Qp\BsFileControl $btnFolder;
+    protected Qp\BsFileControl $btnGroup1;
+    protected Qp\BsFileControl $btnGroup2;
 
-    protected function formCreate()
+    /**
+     * Initializes file upload controls for single and multiple files, as well as folder uploads.
+     * Configures file controls with different styles, icons, and behaviors such as allowing multiple file uploads or folder uploads.
+     *
+     * @return void
+     * @throws Caller
+     */
+    protected function formCreate(): void
     {
-        $this->btnSingle = new Q\Plugin\BsFileControl($this);
+        $this->btnSingle = new Qp\BsFileControl($this);
         $this->btnSingle->Text = t('Single upload');
         $this->btnSingle->Glyph = 'fa fa-upload';
         $this->btnSingle->CssClass = 'btn btn-default  fileinput-button';
 
-        $this->btnMultiple = new Q\Plugin\BsFileControl($this);
+        $this->btnMultiple = new Qp\BsFileControl($this);
         $this->btnMultiple->Text = t('File upload');
         $this->btnMultiple->Glyph = 'fa fa-upload';
         $this->btnMultiple->CssClass = 'btn btn-success fileinput-button';
         $this->btnMultiple->Multiple = true;
 
-        $this->btnFolder = new Q\Plugin\BsFileControl($this);
+        $this->btnFolder = new Qp\BsFileControl($this);
         $this->btnFolder->Text = t('Folder upload');
         $this->btnFolder->Glyph = 'fa fa-folder-open';
         $this->btnFolder->CssClass = 'btn btn-primary fileinput-button';
         $this->btnFolder->Multiple = true;
         $this->btnFolder->Folder = true;
 
-        $this->btnGroup1 = new Q\Plugin\BsFileControl($this);
+        $this->btnGroup1 = new Qp\BsFileControl($this);
         $this->btnGroup1->Text = t('File upload');
         //$this->btnGroup1->Glyph = 'fa fa-upload';
         $this->btnGroup1->CssClass = 'btn btn-orange fileinput-button';
         $this->btnGroup1->Multiple = true;
         $this->btnGroup1->UseWrapper = false;
 
-        $this->btnGroup2 = new Q\Plugin\BsFileControl($this);
+        $this->btnGroup2 = new Qp\BsFileControl($this);
         $this->btnGroup2->Text = t('Folder upload');
         //$this->btnGroup2->Glyph = 'fa fa fa-folder-open';
         $this->btnGroup2->CssClass = 'btn btn-orange fileinput-button';
@@ -51,4 +59,4 @@ class ExamplesForm extends Form
     }
 
 }
-ExamplesForm::Run('ExamplesForm');
+ExamplesForm::run('ExamplesForm');
